# DOCS/API.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SARVAL API",
    "version": "1.0.0",
    "description": "API del planificador de descargas en silos de SARVAL.\n\nToda la API está protegida con JWT bearer salvo `POST /api/auth/login`. El motor de simulación se recalcula **automáticamente** en cualquier endpoint que cambie el plan (subir Excel, crear/editar/borrar viaje, parada, box, lectura de nivel, franja de productividad o parámetros de tolva). El único recálculo manual con notificación de webhooks es `POST /api/planning/recalculate`.\n\n**Concepto \"plan activo\"**: la mayoría de endpoints aceptan `?plan_id=…`. Si no se pasa, se usa el plan vigente (el de la semana en curso).\n\n**Semana operativa**: Lunes 06:00 → Sábado 22:00 (consumo continuo, también de noche).\n\nVer `DOCS/API.md` en el repo para los flujos típicos."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local"
    },
    {
      "url": "https://sarval.example.com",
      "description": "Producción (ajustar)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "Trip": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1234
          },
          "plan_id": {
            "type": "integer",
            "example": 28
          },
          "trip_number": {
            "type": "string",
            "description": "Matrícula o identificador del viaje (único por plan)",
            "example": "5678ABC"
          },
          "day": {
            "type": "string",
            "enum": [
              "Lunes",
              "Martes",
              "Miércoles",
              "Jueves",
              "Viernes",
              "Sábado"
            ],
            "example": "Martes"
          },
          "scheduled_time": {
            "type": "string",
            "example": "09:30",
            "description": "HH:MM"
          },
          "supplier": {
            "type": "string",
            "example": "TRANSPORTES X"
          },
          "producto": {
            "type": "string",
            "example": "TRIGO"
          },
          "tons": {
            "type": "number",
            "example": 27.5
          },
          "is_critical": {
            "type": "boolean",
            "example": false
          },
          "is_extra": {
            "type": "boolean",
            "description": "true si fue añadido manualmente (no venía del Excel)",
            "example": false
          },
          "delay_h": {
            "type": "number",
            "nullable": true,
            "example": 0
          },
          "new_time": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "status": {
            "type": "string",
            "nullable": true,
            "example": "pendiente"
          },
          "tolva_id": {
            "type": "integer",
            "example": 1
          },
          "tolva_numero": {
            "type": "integer",
            "example": 1
          },
          "tolva_nombre": {
            "type": "string",
            "example": "Tolva 1"
          }
        }
      },
...
    },
    "responses": {
...
      "NotFound": {
        "description": "Recurso no encontrado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Petición inválida.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/trips": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Listar viajes del plan",
        "parameters": [
          {
            "name": "plan_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Plan a consultar. Si se omite se usa el plan activo (vigente)."
          },
          {
            "name": "tolva_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Filtra por tolva."
          },
          {
            "name": "day",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "Lunes",
                "Martes",
                "Miércoles",
                "Jueves",
                "Viernes",
                "Sábado"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de viajes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Trip"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{id}": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Obtener un viaje",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "trips"
        ],
        "summary": "Editar viaje (recalcula plan)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Solo los campos a actualizar.",
                "properties": {
                  "day": {
                    "type": "string"
                  },
                  "scheduled_time": {
                    "type": "string",
                    "example": "10:30"
                  },
                  "supplier": {
                    "type": "string"
                  },
                  "producto": {
                    "type": "string"
                  },
                  "tons": {
                    "type": "number"
                  },
                  "is_critical": {
                    "type": "boolean"
                  },
                  "delay_h": {
                    "type": "number"
                  },
                  "new_time": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "tolva_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "trips"
        ],
        "summary": "Eliminar viaje (recalcula plan)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/trips/extra": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Añadir viaje extra (recalcula plan)",
        "parameters": [
          {
            "name": "plan_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Plan a consultar. Si se omite se usa el plan activo (vigente)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "trip_number",
                  "day",
                  "scheduled_time",
                  "supplier",
                  "producto",
                  "tons",
                  "tolva_id"
                ],
                "properties": {
                  "trip_number": {
                    "type": "string",
                    "example": "9999XYZ"
                  },
                  "day": {
                    "type": "string",
                    "example": "Miércoles"
                  },
                  "scheduled_time": {
                    "type": "string",
                    "example": "14:00"
                  },
                  "supplier": {
                    "type": "string",
                    "example": "TRANSPORTES X"
                  },
                  "producto": {
                    "type": "string",
                    "example": "MAÍZ"
                  },
                  "tons": {
                    "type": "number",
                    "example": 28
                  },
                  "is_critical": {
                    "type": "boolean",
                    "example": false
                  },
                  "tolva_id": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "description": "Ya existe un viaje con esa matrícula en la semana."
          }
        }
      }
    },
    "/api/trips/{id}/divert": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Desviar viaje a otra tolva (recalcula ambas)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tolva_id"
                ],
                "properties": {
                  "tolva_id": {
                    "type": "integer",
                    "description": "Tolva destino",
                    "example": 2
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Desviado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "trip_id": {
                      "type": "integer"
                    },
                    "from_tolva_id": {
                      "type": "integer"
                    },
                    "to_tolva_id": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}